Accepts integers for Position/Direction/Point

+++ km3pipe/dataclasses.pyx
@@ -25,7 +25,7 @@ __all__ = ('Point', 'Position', 'Direction', 'HitSeries', 'HitSeriesA', 'Hit',
 point_dt = np.dtype([('x', float), ('y', float), ('z', float)])
 
 def Point(vector, as_recarray=True):
-    vector = np.array(vector)
+    vector = np.array(vector, dtype=np.float)
     if as_recarray:
         return vector.view(point_dt, np.recarray)
     else:
